Minor adjustment to creds.app enum settings
