+++ airsonic-main/src/main/java/org/airsonic/player/service/AudioScrobblerService.java
@@ -23,12 +23,8 @@ import org.airsonic.player.domain.MediaFile;
 import org.airsonic.player.domain.UserCredential;
 import org.airsonic.player.domain.UserCredential.App;
 import org.airsonic.player.domain.UserSettings;
-import org.airsonic.player.security.GlobalSecurityConfig;
-import org.airsonic.player.security.PasswordDecoder;
 import org.airsonic.player.service.scrobbler.LastFMScrobbler;
 import org.airsonic.player.service.scrobbler.ListenBrainzScrobbler;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -42,24 +38,12 @@ import java.util.Map;
  */
 @Service
 public class AudioScrobblerService {
-    private static final Logger LOG = LoggerFactory.getLogger(AudioScrobblerService.class);
-
     private LastFMScrobbler lastFMScrobbler;
     private ListenBrainzScrobbler listenBrainzScrobbler;
     @Autowired
     private SettingsService settingsService;
     @Autowired
     private SecurityService securityService;
-
-    private static final String decode(UserCredential uc) {
-        PasswordDecoder decoder = (PasswordDecoder) GlobalSecurityConfig.ENCODERS.get(uc.getEncoder());
-        try {
-            return decoder.decode(uc.getCredential());
-        } catch (Exception e) {
-            LOG.warn("Could not decode credentials for user {}, app {}", uc.getUsername(), uc.getApp(), e);
-            return null;
-        }
-    }
 
     /**
      * Registers the given media file at audio scrobble service.
@@ -94,7 +78,7 @@ public class AudioScrobblerService {
             if (lastFMScrobbler == null) {
                 lastFMScrobbler = new LastFMScrobbler();
             }
-            String decoded = decode(cred);
+            String decoded = SecurityService.decodeCredentials(cred);
             if (decoded != null) {
                 lastFMScrobbler.register(mediaFile, cred.getAppUsername(), decoded, submission, time);
             }
@@ -105,7 +89,7 @@ public class AudioScrobblerService {
             if (listenBrainzScrobbler == null) {
                 listenBrainzScrobbler = new ListenBrainzScrobbler();
             }
-            String decoded = decode(cred);
+            String decoded = SecurityService.decodeCredentials(cred);
             if (decoded != null) {
                 listenBrainzScrobbler.register(mediaFile, userSettings.getListenBrainzUrl(), decoded, submission, time);
             }
